pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "whispaste"
version = "0.1.0"
description = "Simple voice-to-paste tool"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "sounddevice>=0.5.0",
    "numpy>=1.24.0",
    # We keep pyperclip as a fallback, though we prefer system commands
    "pyperclip>=1.8.0",
]

[project.scripts]
whispaste = "whispaste.__main__:main"

[tool.setuptools]
packages = ["whispaste"]
